create background pattern component
