+++ src/components/index.ts
@@ -11,3 +11,4 @@ export { default as ChannelGrid } from './ChannelGrid'
 export { default as ViewErrorFallback } from './ViewErrorFallback'
 export { default as ErrorFallback } from './ErrorFallback'
 export { default as ChannelLink } from './ChannelLink'
+export { default as BackgroundPattern } from './BackgroundPattern'
